+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 
+## [0.3.0] ?
+
+- Introduce CMakeTools' build kit -specific `bin` directory command and PATH command for use within `launch.json`
+
+- Documentation update
+
 ## [0.2.0] 2021-2-6
 
-- Introduce a few CMakeTools build kit -specific commands `cmake.buildkit.*.exe` to enable the use of kit-specific CMake & generator tools instead of the generic MSYS2-specific ones.
+- Introduce a few CMakeTools' build kit -specific commands `cmake.buildkit.*.exe` to enable the use of kit-specific CMake & generator tools instead of the generic MSYS2-specific ones.
 
 - Per user settings template update